+++ public/assets/tinymce/plugins/uploadimage/langs/nb.js
@@ -1,0 +1,3 @@
+tinyMCE.addI18n('nb.uploadimage', {
+  desc: 'Sett inn et bilde fra datamaskinen'
+});